Build types explicitly via tsconfig.types.json
Update packages and fix unknown types
Remove PNPM cache
Add previously ignored file
Fix default types
