+++ .gitignore
@@ -1,6 +1,7 @@
 node_modules/
 dist/
 types/
+!src/types/
 coverage/
 
 # IDE
