+++ package.json
@@ -16,16 +16,18 @@
   "exports": {
     ".": {
       "types": "./dist/index.d.ts",
-      "default": "./dist/index.js"
+      "import": "./dist/index.js",
+      "require": "./dist/index.js"
     },
     "./types": {
       "types": "./types/index.d.ts",
-      "default": "./types/index.js"
+      "import": "./types/index.js",
+      "require": "./types/index.js"
     }
   },
-  "types": "./types/index.d.ts",
+  "types": "./dist/index.d.ts",
   "scripts": {
-    "build": "rimraf ./types ./dist && tsc",
+    "build": "rimraf ./types ./dist && tsc --project ./tsconfig.types.json && tsc",
     "lint": "eslint 'src/**/*.{js,ts}'",
     "test": "jest",
     "test:coverage": "jest; open ./coverage/lcov-report/index.html",
@@ -33,36 +35,36 @@
     "prepare": "pnpm i rimraf && pnpm build"
   },
   "dependencies": {
-    "fast-xml-parser": "^4.1.3",
+    "fast-xml-parser": "^4.2.2",
     "graphql": "^16.6.0",
-    "graphql-request": "^5.1.0",
+    "graphql-request": "^5.2.0",
     "lodash.get": "^4.4.2",
     "lodash.has": "^4.5.2",
     "lodash.kebabcase": "^4.1.1",
     "lodash.property": "^4.4.2",
     "lodash.set": "^4.3.2",
-    "type-fest": "^3.6.1",
+    "type-fest": "^3.8.0",
     "zod": "^3.21.4"
   },
   "devDependencies": {
-    "@types/jest": "^29.4.0",
+    "@types/jest": "^29.5.1",
     "@types/lodash.get": "^4.4.7",
     "@types/lodash.has": "^4.5.7",
     "@types/lodash.kebabcase": "^4.1.7",
     "@types/lodash.property": "^4.4.7",
     "@types/lodash.set": "^4.3.7",
-    "@types/node": "^18.11.11",
-    "@typescript-eslint/eslint-plugin": "^5.45.1",
-    "@typescript-eslint/parser": "^5.45.1",
-    "esbuild": "^0.17.10",
+    "@types/node": "^18.16.0",
+    "@typescript-eslint/eslint-plugin": "^5.59.1",
+    "@typescript-eslint/parser": "^5.59.1",
+    "esbuild": "^0.17.18",
     "esbuild-jest": "^0.5.0",
-    "eslint": "^8.29.0",
-    "jest": "^29.4.3",
-    "prettier": "^2.8.0",
+    "eslint": "^8.39.0",
+    "jest": "^29.5.0",
+    "prettier": "^2.8.8",
     "rimraf": "^3.0.2",
-    "ts-jest": "^29.0.5",
+    "ts-jest": "^29.1.0",
     "ts-node": "^10.9.1",
-    "typescript": "^4.9.3"
+    "typescript": "^4.9.5"
   },
   "engines": {
     "node": "^16",
