+++ src/types/flow.ts
@@ -1,0 +1,34 @@
+import { ComponentType } from "./component";
+import type { Value } from "./data";
+
+export type NodeId = string;
+
+export type Edges = Array<NodeId>;
+
+export interface Node {
+  id?: string;
+  type?: ComponentType;
+  edges?: Edges;
+  data?: Record<NodeId, Value>;
+}
+
+export type FlowGraph = {
+  _root: {
+    edges: Edges;
+  };
+  [key: string]: Node;
+};
+
+export interface IndexedNode extends Node {
+  id: string;
+  parentId: string | null; // null if it is the first node
+}
+
+export type OrderedFlow = Array<IndexedNode>;
+
+export interface NormalizedNode extends IndexedNode {
+  sectionId?: string;
+  rootNodeId: string;
+}
+
+export type NormalizedFlow = Array<NormalizedNode>;
