+++ tsconfig.json
@@ -13,7 +13,6 @@
     "target": "es2021",
     "rootDir": "./src",
     "declaration": true,
-    "declarationDir": "./types",
     "outDir": "dist"
   },
   "include": ["src/index.ts"]
