+++ tsconfig.types.json
@@ -11,8 +11,10 @@
     "skipLibCheck": true,
     "strict": true,
     "target": "es2021",
-    "rootDir": "./src",
-    "noEmit": true
+    "rootDir": "./src/types",
+    "declaration": true,
+    "declarationDir": "types",
+    "outDir": "types"
   },
-  "include": [ "src" ]
+  "include": ["src/types/index.ts"]
 }
